lai jupyter component
