--- tests/README.md
@@ -1,17 +1,0 @@
-# Run tests
-
-To run the tests:
-
-```bash
-# go to your component folder
-cd lit_jupyter
-
-# go to tests folder
-cd tests
-
-# install testing deps
-pip install -r requirements.txt
-
-# run tests
-pytest .
-```
