+++ tests/requirements.txt
@@ -6,3 +6,4 @@ pytest-flake8
 flake8
 check-manifest
 twine==1.13.0
+jupyterlab==3.4.3
